implemented close garage feature, and updated ticketPrice data type

+++ src/com/company/Parkable.java
@@ -1,8 +1,8 @@
 package com.company;
 
 public interface Parkable {
-    int PARKING_FEE = 500;
-    public int getParkingFee();
+    double PARKING_FEE = 500;
+    public double getParkingFee();
     public void checkIn();
     public void checkOut();
     public void lostTicket();
